Warehouse 672023006/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.1
pymongo==4.7.2
PyJWT==2.9.0
MarkupSafe==2.1.3
Werkzeug==2.3.7
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
bleach==6.1.0
flask-limiter==3.5.0
gunicorn==21.2.0
python-dotenv==1.0.0
dnspython==2.4.2
webencodings==0.5.1
six==1.16.0
eventlet==0.35.0